Exclude all transitive dependencies in maven package-tests
This change should allow us to test that the GATK jar has been
correctly packaged at release time, by ensuring that only the
packaged jar + a few test-related dependencies are on the classpath
when tests are run.

Note that we still need to actually test that this works as intended
before we can make this live in the Bamboo release plan.