Provided AlleleBalanceBySample annotation in GVCFs
Refactored AlleleBalanceBySample to use a PerReadAlleleLikelihoodMap as
well as a AlignmentContext.  Additionally, accounted for the <NON_REF>
allele in the GVCF as called by the HaplotypeCaller.  This allows the
"AB" annotation in the FORMAT field to be carried though the
GenotypeGVCFs utitlity, preventing the need for a VariantAnnotator call
on the combined VCF, which requires re-reading all of the supporting BAM
files, which is both time and memory intensive.

Note that most of the code was copied from the DepthPerAlleleBySample
class; It would probably be good to refactor the common code into a
shared base class or shared utility class.

Signed-off-by: John Wallace <jrw32@psu.edu>