Fixed NullPointerException in GenotypeGVCFs
Fixed a NullPointerException when using the GenotypeGVCFs utility with
the AlleleBalance annotation.  Because GenotypeGVCFs passes a null
Mapping of samples to StratifiedContexts, AlleleBalance was throwing a
NullPointerException on line 66, checking for a zero-sized map before
checking for a null map.

Modified AlleleBalance to not require a StratifiedContext map for
calculating the allele balance.  Instead, now use the AD field in the
gentoype to calculate the allele balance.  If no AD field is provided,
fail gracefully and do not include that sample in the AlleleBalance
calculations.

Signed-off-by: John Wallace <jrw32@psu.edu>