Now AssessNA12878 can handle different input ploidy for omniploidy assessment where we may run HC as if it was a haploid, tetraploid (and so forth) organism.
For example, when the input is Haploid it is considered ok to have a FN if the actual genotype is 0/1 as there is 50% chance to not call it at all.

Also it considers that the genotype call is concordant as long as the AC is as close as it can be to the 50% percent given the ploidy. So for a 0/1 true call is it ok
to have a 0 or 1 call in haploids and also 0/0/1/1 in tetraploid, and also 0/0/1 or 0/1/1 with triploid input, but it is not a 0/0/0/1 in tetraploids or 0/0/0/0/1/1 with hextaploid input.

Story:

  http://www.pivotaltracker.com/story/show/72090992

Changes:

  AssessNA12878 has a new argument (-ploidy / --inputPloidy) to indicate the expected ploidy of the input.
  By default this is the obvious choice of 2 as NA12878 is human.

  In the input has calls with a different ploidy it will complain with an user exception.

  Also some refactoring has been done to make the code a bit more concise in some parts.